requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
cffi==1.12.0
constantly==15.1.0
cryptography==2.5
cssselect==1.0.3
Django==2.1.7
et-xmlfile==1.0.1
geographiclib==1.49
geopy==1.18.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4
lxml==4.3.1
openpyxl==2.6.0
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyhumps==1.2.2
pyOpenSSL==19.0.0
pytz==2018.9
queuelib==1.5.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.12.0
Twisted==18.9.0
w3lib==1.20.0
zope.interface==4.6.0
